# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  docs.BatchDetailEntryResponse:
    properties:
      host_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      reason:
        example: 'skip: host knowledge is sufficient'
        type: string
      scan_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      stage:
        example: os_detection
        type: string
    type: object
  docs.BatchResultResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/docs.BatchDetailEntryResponse'
        type: array
      queued:
        example: 15
        type: integer
      skipped:
        example: 51
        type: integer
    type: object
...
  docs.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      request_id:
        example: req-123
        type: string
      timestamp:
        type: string
    type: object
...
  docs.ProfileRecommendationResponse:
    properties:
      action:
        example: port_expansion
        type: string
      host_count:
        example: 42
        type: integer
      os_family:
        example: Linux
        type: string
      profile_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      profile_name:
        example: Linux Standard
        type: string
    type: object
...
  docs.RefreshIdentityResponse:
    properties:
      host_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      queued:
        example: true
        type: boolean
      scan_id:
        example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        type: string
    type: object
...
  docs.ScanStageResponse:
    properties:
      os_detection:
        example: true
        type: boolean
      ports:
        example: 1-1024
        type: string
      profile_id:
        example: linux-common
        type: string
      reason:
        example: no OS information recorded
        type: string
      scan_type:
        example: syn
        type: string
      stage:
        enum:
        - os_detection
        - port_expansion
        - service_scan
        - refresh
        - skip
        example: os_detection
        type: string
    type: object
...
  docs.SuggestionGroupResponse:
    properties:
      action:
        example: os_detection
        type: string
      count:
        example: 12
        type: integer
      description:
        example: Hosts with no OS information
        type: string
    type: object
  docs.SuggestionSummaryResponse:
    properties:
      generated_at:
        type: string
      no_os_info:
        $ref: '#/definitions/docs.SuggestionGroupResponse'
      no_ports:
        $ref: '#/definitions/docs.SuggestionGroupResponse'
      no_services:
        $ref: '#/definitions/docs.SuggestionGroupResponse'
      stale:
        $ref: '#/definitions/docs.SuggestionGroupResponse'
      total_hosts:
        example: 1082
        type: integer
      well_known:
        $ref: '#/definitions/docs.SuggestionGroupResponse'
    type: object
  docs.TriggerBatchRequest:
    properties:
      host_ids:
        example:
        - 550e8400-e29b-41d4-a716-446655440000
        items:
          type: string
        type: array
      limit:
        example: 50
        type: integer
      network_cidr:
        example: 192.168.1.0/24
        type: string
      os_family:
        example: Linux
        type: string
      stage:
        enum:
        - os_detection
        - port_expansion
        - service_scan
        - refresh
        - skip
        example: os_detection
        type: string
    type: object
  docs.TriggerHostResponse:
    properties:
      host_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      message:
        example: no scan needed — host knowledge is sufficient
        type: string
      queued:
        example: true
        type: boolean
      scan_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Scanorama Support
    url: https://github.com/anstrom/scanorama
  description: |-
    Enterprise-grade network scanning and discovery service with automated reconnaissance capabilities

    ## Features
    - **Advanced Scanning Engine**: Multiple scan types (connect, SYN, version, comprehensive, aggressive, stealth)
    - **Enterprise Reliability**: Race condition-free worker pools with graceful shutdown
    - **Database Integration**: PostgreSQL persistence with automatic migrations and transaction support
    - **Real-time Updates**: WebSocket support for live scan progress and results
    - **Comprehensive API**: RESTful endpoints with full CRUD operations
    - **Monitoring & Observability**: Built-in metrics, structured logging, and health checks
    - **Security**: Vulnerability scanning integration and secure error handling
    - **Scheduling**: Automated scan jobs with cron-like scheduling
    - **High Performance**: Concurrent processing with configurable rate limiting

    ## Quality Assurance
    - **Test Coverage**: >90% coverage on core packages with comprehensive integration tests
    - **Security**: Zero known vulnerabilities with automated security scanning
    - **Code Quality**: Zero linting issues with automated quality checks

    ## Authentication
    Most endpoints require API key authentication. Include your API key in the `X-API-Key` header.
...
  license:
    name: MIT
...
  version: 0.7.0
paths:
...
  /smart-scan/hosts/{id}/refresh-identity:
    post:
      description: Queues an identity_enrichment scan for a host, probing mDNS/SNMP/DNS/TLS
        surfaces so post-scan enrichment can fill in name signals. Runs even when
        the host already has a usable name.
      operationId: refreshHostIdentity
      parameters:
      - description: Host UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/docs.RefreshIdentityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Refresh host identity
      tags:
      - Smart Scan
  /smart-scan/hosts/{id}/stage:
    get:
      description: Returns the recommended next scan stage for the specified host.
      operationId: evaluateHostStage
      parameters:
      - description: Host UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.ScanStageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Evaluate next scan stage for a host
      tags:
      - Smart Scan
  /smart-scan/hosts/{id}/trigger:
    post:
      description: Evaluates the host's knowledge gaps and queues the appropriate
        scan.
      operationId: triggerSmartScan
      parameters:
      - description: Host UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: No action needed
          schema:
            $ref: '#/definitions/docs.TriggerHostResponse'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/docs.TriggerHostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Trigger Smart Scan for a host
      tags:
      - Smart Scan
  /smart-scan/profile-recommendations:
    get:
      description: Returns profile suggestions grouped by detected OS family.
      operationId: getProfileRecommendations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/docs.ProfileRecommendationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profile recommendations
      tags:
      - smart-scan
  /smart-scan/suggestions:
    get:
      description: Returns fleet-wide counts of hosts in each knowledge-gap category.
      operationId: getSmartScanSuggestions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.SuggestionSummaryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Smart Scan suggestions
      tags:
      - Smart Scan
  /smart-scan/trigger-batch:
    post:
      consumes:
      - application/json
      description: Queues smart scans for all eligible hosts matching the filter.
      operationId: triggerSmartScanBatch
      parameters:
      - description: Batch filter
        in: body
        name: body
        schema:
          $ref: '#/definitions/docs.TriggerBatchRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/docs.BatchResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Batch trigger Smart Scan
      tags:
      - Smart Scan
...
security:
- ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    description: API key for authentication
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"